+++ packages/plugins/Attributions/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Chore: Upgrade `ol` from `^9.2.4` to `^10.3.1`.
+
 ## 1.3.0
 
 - Feature: Add new option `'footer'` to configuration parameter `renderType` that changes the attributions to be displayed as a small version of the information box that is always visible.